0.9.3 Cleanup and simplification removed lighting and anything related to it remove datatable removed minimap module (might come back) move priority queue to datastructures removed shader data implemented gamesystem benchmarking improved gamesystem API simplified benchmark code updated road map added absolute scaling mode for UIObjects added keybinds to KeyBind many other cleanups using modern types

+++ include/magique/fwd.hpp
@@ -7,7 +7,7 @@
 
 //================= USER DEFINED =================//
 
-enum class GameState : int;     // Identifies different game states
+enum class GameState : uint8_t; // Identifies different game states
 enum EntityType : uint16_t;     // Identifies entity types       - uses uint16_t to be small for networking
 enum class MapID : uint8_t;     // Identifies maps               - uses uint8_t  to be small for networking
 enum class HandleID : int;      // Identifies handles
@@ -149,7 +149,7 @@ namespace entt
 
 namespace glz
 {
-    template<typename T>
+    template <typename T>
     struct meta;
 }
